Add database 2

+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/dao/FilmDbStorage.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.excepton.film.FilmDoesntExistException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
+import ru.yandex.practicum.filmorate.service.UserService;
 import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
 import ru.yandex.practicum.filmorate.storage.genre.GenreMapper;
 
@@ -15,17 +16,14 @@ import java.util.*;
 public class FilmDbStorage implements FilmStorage {
 
     private final JdbcTemplate jdbcTemplate;
-    private final String filmSql = "SELECT f.*, m.mpa_id, m.name as mpa_name " +
+    private final UserService userService;
+    private static final String FILM_SQL = "SELECT f.*, m.mpa_id, m.name as mpa_name " +
             "FROM films f " +
             "INNER JOIN mpa_ratings m ON f.mpa_id = m.mpa_id ";
-//    private final String filmsSql = "SELECT f.film_id, f.name, f.description, f.releaseDate, f.duration, m.mpa_id, m.name AS mpa_name, g.genre_id, g.genre_name AS genre_name " +
-//            "FROM films f " +
-//            "INNER JOIN mpa_ratings m ON f.mpa_id = m.mpa_id " +
-//            "LEFT JOIN film_genres fg ON f.film_id = fg.film_id " +
-//            "LEFT JOIN genres g ON fg.genre_id = g.genre_id ";
 
-    public FilmDbStorage(JdbcTemplate jdbcTemplate) {
+    public FilmDbStorage(JdbcTemplate jdbcTemplate, UserService userService) {
         this.jdbcTemplate = jdbcTemplate;
+        this.userService = userService;
     }
 
     @Override
@@ -33,7 +31,7 @@ public class FilmDbStorage implements FilmStorage {
         SimpleJdbcInsert simpleJdbcInsert = new SimpleJdbcInsert(jdbcTemplate)
                 .withTableName("Films")
                 .usingGeneratedKeyColumns("film_id")
-                .usingColumns("name", "description", "releaseDate", "duration", "mpa_id");;
+                .usingColumns("name", "description", "releaseDate", "duration", "mpa_id");
 
         Map<String, Object> parameters = new HashMap<>();
         parameters.put("name", film.getName());
@@ -44,6 +42,14 @@ public class FilmDbStorage implements FilmStorage {
 
         Number filmId = simpleJdbcInsert.executeAndReturnKey(parameters);
         film.setId(filmId.intValue());
+
+        if (film.getGenres() != null) {
+            // add genres
+            for (Genre genre : film.getGenres()) {
+                jdbcTemplate.update("INSERT INTO film_genres (film_id, genre_id) VALUES (?, ?)",
+                        film.getId(), genre.getId());
+            }
+        }
 
         return film;
     }
@@ -58,36 +64,26 @@ public class FilmDbStorage implements FilmStorage {
             throw new FilmDoesntExistException("Film with id: " + film.getId() + " does not exist");
         }
 
-        // Удаление старых связей фильма с жанрами
+        // removing old genres from a movie
         jdbcTemplate.update("DELETE FROM film_genres WHERE film_id=?", film.getId());
 
-        if (film.getGenres() != null){
-            // Вставка новых связей фильма с жанрами
+        if (film.getGenres() != null) {
+            // inserting new genres into the film
             for (Genre genre : film.getGenres()) {
                 jdbcTemplate.update("INSERT INTO film_genres (film_id, genre_id) VALUES (?, ?)", film.getId(), genre.getId());
             }
         }
 
-        // Обновление основных данных фильма
+        // updating basic movie data
         jdbcTemplate.update("UPDATE films SET name=?, description=?, releaseDate=?, duration=?, mpa_id=? WHERE film_id=?",
                 film.getName(), film.getDescription(), film.getReleaseDate(), film.getDuration(), film.getMpa().getId(), film.getId());
 
         return film;
     }
 
-//    @Override
-//    public List<Film> getFilms() {
-//        String sql = "SELECT f.film_id, f.name, f.description, f.releaseDate, f.duration, m.mpa_id, m.name AS mpa_name, g.genre_id, g.genre_name AS genre_name " +
-//                "FROM films f " +
-//                "INNER JOIN mpa_ratings m ON f.mpa_id = m.mpa_id " +
-//                "LEFT JOIN film_genres fg ON f.film_id = fg.film_id " +
-//                "LEFT JOIN genres g ON fg.genre_id = g.genre_id ";
-//        return jdbcTemplate.query(sql, new FilmMapper());
-//    }
-
     @Override
     public List<Film> getFilms() {
-        List<Film> films = jdbcTemplate.query(filmSql, new FilmMapper());
+        List<Film> films = jdbcTemplate.query(FILM_SQL, new FilmMapper());
         //add genres to films
         films.forEach(this::addGenresToFilm);
 
@@ -96,7 +92,7 @@ public class FilmDbStorage implements FilmStorage {
 
     @Override
     public Film getFilmById(int id) {
-        String sql = filmSql +
+        String sql = FILM_SQL +
                 "WHERE f.film_id=?";
 
         Film film = jdbcTemplate.query(sql, new FilmMapper(), id)
@@ -110,15 +106,21 @@ public class FilmDbStorage implements FilmStorage {
 
     //likes
     @Override
-    public void addLike(int id, int userId) {
-        jdbcTemplate.update("INSERT INTO Film_likes (film_id, user_id) VALUES (?, ?)",
-                id, userId);
+    public void addLike(int filmId, int userId) {
+        //check that the film and the user exist
+        getFilmById(filmId);
+        userService.getUser(userId);
+
+        jdbcTemplate.update("INSERT INTO Film_likes (film_id, user_id) VALUES (?, ?)", filmId, userId);
     }
 
     @Override
-    public void removeLike(int id, int userId) {
-        jdbcTemplate.update("DELETE FROM Film_likes WHERE film_id=? and user_id=?",
-                id, userId);
+    public void removeLike(int filmId, int userId) {
+        //check that the film and the user exist
+        getFilmById(filmId);
+        userService.getUser(userId);
+
+        jdbcTemplate.update("DELETE FROM Film_likes WHERE film_id=? and user_id=?", filmId, userId);
     }
 
 
@@ -128,17 +130,19 @@ public class FilmDbStorage implements FilmStorage {
             count = 10; // If count is not set or is invalid, return the first 10 movies
         }
 
-        String sql = "SELECT f.* " +
-                "FROM films f " +
-                "JOIN film_likes fl ON f.film_id = fl.film_id " +
+        String sql = FILM_SQL +
+                "LEFT JOIN film_likes fl ON f.film_id = fl.film_id " +
                 "GROUP BY f.film_id " +
                 "ORDER BY COUNT(fl.film_id) DESC " +
                 "LIMIT ?";
 
-        return jdbcTemplate.query(sql, new FilmMapper(), count);
+        List<Film> films = jdbcTemplate.query(sql, new FilmMapper(), count);
+        //add genres to films
+        films.forEach(this::addGenresToFilm);
+        return films;
     }
 
-    private void addGenresToFilm(Film film){
+    private void addGenresToFilm(Film film) {
         String sql = "SELECT g.genre_id, g.genre_name " +
                 "FROM films f " +
                 "LEFT  JOIN film_genres fg ON f.film_id = fg.film_id " +
@@ -148,26 +152,12 @@ public class FilmDbStorage implements FilmStorage {
         List<Genre> filmGenres = jdbcTemplate.query(sql, new GenreMapper(), film.getId());
 
         //checking if film have no genres
-        if (filmGenres.contains(Genre.builder().id(0).name(null).build())){
+        if (filmGenres.contains(Genre.builder().id(0).name(null).build())) {
             film.setGenres(new TreeSet<>());
             return;
         }
 
         film.setGenres(new TreeSet<>(filmGenres));
     }
-
-//    private Film rsToFilm(ResultSet resultSet) throws SQLException {
-//        return Film.builder()
-//                .id(resultSet.getInt("film_id"))
-//                .name(resultSet.getString("name"))
-//                .description(resultSet.getString("description"))
-//                .releaseDate(resultSet.getDate("release_date").toLocalDate())
-//                .duration(resultSet.getInt("duration"))
-//                .mpa(Mpa.builder()
-//                        .id(resultSet.getInt("MPA_RATINGS.mpa_id"))
-//                        .name(resultSet.getString("MPA_RATINGS.name"))
-//                        .build())
-//                .build();
-//    }
 
 }
